[metrics] Changing variable types. Adjusting the license.

+++ lib/src/system/sys_metrics_processor.cc
@@ -1,28 +1,19 @@
 /**
- * Copyright 2013-2021 Software Radio Systems Limited
  *
- * This file is part of srsLTE.
+ * \section COPYRIGHT
  *
- * srsLTE is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of
- * the License, or (at your option) any later version.
+ * Copyright 2013-2020 Software Radio Systems Limited
  *
- * srsLTE is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * A copy of the GNU Affero General Public License can be found in
- * the LICENSE file in the top-level directory of this distribution
- * and at http://www.gnu.org/licenses/.
+ * By using this file, you agree to the terms and conditions set
+ * forth in the LICENSE file which can be found at the top level of
+ * the distribution.
  *
  */
 
 #include "srslte/system/sys_metrics_processor.h"
-#include "sys/sysinfo.h"
 #include <fstream>
 #include <sstream>
+#include <sys/sysinfo.h>
 #include <unistd.h>
 
 using namespace srslte;
@@ -51,7 +42,7 @@ sys_metrics_processor::proc_stats_info::proc_stats_info()
 sys_metrics_t sys_metrics_processor::get_metrics()
 {
   auto     current_time = std::chrono::steady_clock::now();
-  unsigned measure_interval_ms =
+  uint32_t measure_interval_ms =
       std::chrono::duration_cast<std::chrono::milliseconds>(current_time - last_query_time).count();
 
   // The time elapsed between 2 measures must be greater that 100 milliseconds.
@@ -89,7 +80,7 @@ float sys_metrics_processor::calculate_cpu_usage(const proc_stats_info& current_
     return 0.f;
   }
 
-  static const unsigned cpu_count        = ::sysconf(_SC_NPROCESSORS_CONF);
+  static const uint32_t cpu_count        = ::sysconf(_SC_NPROCESSORS_CONF);
   static const float    ticks_per_second = ::sysconf(_SC_CLK_TCK);
 
   return ((current_query.stime + current_query.utime) - (last_query.stime + last_query.utime)) * 100.f /
@@ -97,11 +88,11 @@ float sys_metrics_processor::calculate_cpu_usage(const proc_stats_info& current_
 }
 
 /// Extracts and returns the memory size from the given line.
-static unsigned read_memory_value_from_line(const std::string& line)
+static uint32_t read_memory_value_from_line(const std::string& line)
 {
   std::istringstream reader(line);
   std::string        label, unit;
-  unsigned           value;
+  uint32_t           value;
 
   reader >> label >> value >> unit;
 
@@ -130,12 +121,12 @@ static void calculate_percentage_memory(sys_metrics_t& metrics)
 
   // Total system's memory is in the first line.
   std::getline(file, line);
-  unsigned long long total_mem_kB = read_memory_value_from_line(line);
+  uint32_t total_mem_kB = read_memory_value_from_line(line);
 
   // System's available memory is in the third line.
   std::getline(file, line);
   std::getline(file, line);
-  unsigned long long available_mem_kB = read_memory_value_from_line(line);
+  uint32_t available_mem_kB = read_memory_value_from_line(line);
 
   // Calculate the metrics.
   metrics.process_realmem    = 100.f * (float(metrics.process_realmem_kB) / total_mem_kB);
